+++ project_md03-main/src/main/java/ra/project_md03/model/entity/Order.java
@@ -5,6 +5,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.Date;
 @NoArgsConstructor
 @AllArgsConstructor
@@ -13,7 +14,7 @@ import java.util.Date;
 public class Order {
     private int orderId;
     private User user;
-    private double orderTotal;
+    private BigDecimal orderTotal;
     private int orderStatus = 0;
     private Date order_at;
     private String address;
